requirements.txt
streamlit>=1.30.0
pydantic>=2.0.0
anthropic>=0.40.0
openai>=1.0.0
fpdf2>=2.7.0
